Background the Client Personal ID Lookup XLSX export
  The synchronous xlsx respond_to built the whole crosswalk in-request, which
  times out on large project scopes. Move generation into a DocumentExport job
  and drive the download from the filter form via Stimulus, since this page has
  no intermediate submit for the server to render a query string into.
